+++ R/CohortStats.R
@@ -1,4 +1,4 @@
-# Copyright 2022 Observational Health Data Sciences and Informatics
+# Copyright 2023 Observational Health Data Sciences and Informatics
 #
 # This file is part of CohortGenerator
 #
@@ -146,7 +146,7 @@ getStatsTable <- function(connectionDetails,
     connection <- DatabaseConnector::connect(connectionDetails)
     on.exit(DatabaseConnector::disconnect(connection))
   }
-  
+
   # Force databaseId to NULL when includeDatabaseId is FALSE
   if (!includeDatabaseId) {
     databaseId <- NULL
@@ -160,9 +160,10 @@ getStatsTable <- function(connectionDetails,
     snakeCaseToCamelCase = snakeCaseToCamelCase,
     table = table,
     cohort_database_schema = cohortDatabaseSchema,
-    database_id = ifelse(test = is.null(databaseId), 
-                         yes = "", 
-                         no = databaseId)
+    database_id = ifelse(test = is.null(databaseId),
+      yes = "",
+      no = databaseId
+    )
   )
 
   if (!snakeCaseToCamelCase) {
@@ -209,7 +210,6 @@ getCohortStats <- function(connectionDetails,
                              "cohortCensorStatsTable"
                            ),
                            cohortTableNames = getCohortTableNames()) {
-
   # Names of cohort table names must include output tables
   checkmate::assertNames(names(cohortTableNames), must.include = outputTables)
   # ouput tables strictly the set of allowed tables
@@ -228,9 +228,10 @@ getCohortStats <- function(connectionDetails,
     # The cohortInclusionTable does not hold database
     # specific information so the databaseId
     # should NOT be included.
-    includeDatabaseId = ifelse(test = table != "cohortInclusionTable",
-                               yes = TRUE,
-                               no = FALSE)
+    includeDatabaseId <- ifelse(test = table != "cohortInclusionTable",
+      yes = TRUE,
+      no = FALSE
+    )
     results[[table]] <- getStatsTable(
       connectionDetails = connectionDetails,
       connection = connection,