+++ tests/testthat/setup.R
@@ -1,0 +1,10 @@
+dbmsPlatforms <- c("redshift", "postgresql", "oracle", "sql server")
+connectionDetails <- Eunomia::getEunomiaConnectionDetails()
+outputFolder <- tempfile()
+dir.create(outputFolder)
+withr::defer(
+  {
+    unlink(outputFolder)
+  },
+  testthat::teardown_env()
+)